Add weather forecast move logic from user module
